refactor(log): display packages test lists as table
