+++ libs/bublik/features/session-log/src/lib/v1/log-blocks/log-meta/index.ts
@@ -1,4 +1,3 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /* SPDX-FileCopyrightText: 2021-2023 OKTET Labs Ltd. */
 export * from './log-meta';
-export * from './entity';
